some fixes and internationalization config changes

+++ src/main/java/com/bannerlordonlineplayers/util/validation/HasId.java
@@ -3,16 +3,16 @@ package com.bannerlordonlineplayers.util.validation;
 import org.springframework.util.Assert;
 
 public interface HasId {
-    Integer getId();
+    Long getId();
 
-    void setId(Integer id);
+    void setId(Long id);
 
     default boolean isNew() {
         return getId() == null;
     }
 
     // doesn't work for hibernate lazy proxy
-    default int id() {
+    default Long id() {
         Assert.notNull(getId(), "Entity must has id");
         return getId();
     }
